backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.1.0
Django==4.2.20
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.6
sqlparse==0.5.3
typing-extensions==4.13.1
